Create and remove tasks
createTask() - stores a client, project, task, and description in the
tasks table.
- the tasks table stores a projectID not projectName so createTask
calls the returnProjectID function to get the projectId of the given
task

deleteTask()- deletes a task given its client, project, and task name
removes this row from the tasks table using the removeTasks function.
removeTasks converts projectName into a projectID so the proper row can
be removed.