Create and Delete TimeSheet Entries
createTimeSheet() - passes a username, client name, project name, task
name, time, and time out.
It converts the project name to projectID and taskName to taskID.
then it calls newTimeSheet

deleteTimeSheet()-passes a username, client name, project name, task
name, time, and time out.
It converts the project name to projectID and taskName to taskID.
then it calls removeTimeSheet

the test_demo creates an entry for a timesheet with a mock
developer,client, project and task