Changed ClientPurchases (hoursPurchased) data type
In the ClientPurchases table I changed the hoursPurchased column to a
biginteger, originally it was a DateTime

The reason I did this is because when a client purchases time they will
probably purchase 10 hours, 40 hours, etc.
A dateTime is in the form of ‘YYYY-MM-DD HH-MM-SS’
10 hours = 0000-00-00 10-00-00
40 hours = 0000-00-01 16-00-00
PHP gets confused when it reads 0000-00-00 as the date, it thinks its
the current date. so when php calculates hours purchased and uses 2015
as the year instead of 0 you have a big problem. To fix this problem I
changed the dataType and we are going to store a bigInteger value.
Instead of storing hours we will store seconds.
1s * 3600 = 1 hour

this makes it easier to add purchased hours together and subtract
developer time