Client Reports, Hours Left, Export to excel works
client_reports.php:
- conditional to check if export to excel button has been pressed, if
its been pressed figure out which report was selected and then download
the file as a csv

demo_functions.php:
- formExportToExcel() consumes a client name and a report name. Creates
a form with a submit button for Export to Excel and 2 hidden fields to
send client name and report name

page_functions.php:
- modify clientReport() to call formExportToExcel()

table_functions.php
- printHoursLeftTable() consumes another argument, ‘type’. type should
either be a table or csv. if it is a table it calls printTable(), if
its a csv it calls printCSV()
- cleanData() removes spacing in preperation to export
- cleanTime() formats time from seconds to hours:min:sec in preparation
to export
- printCSV() consumes a query, table headers, and report name. names
the report as the name concatenated with the date. queries the db and
prints the table to a csv